Move DB methods to database file
